Aggiornati i movimenti, questi devono ricevere un importo che sia double e non intero! Successivamente modificato il CMakefile in seguito all'installazione di qt per costruire una piccola iinterfaccina grafica su cui poter inserire i movimenti e creare più conti correnti(file .csv)

+++ Entrata.cpp
@@ -4,7 +4,7 @@
 
 #include "Entrata.h"
 
-Entrata::Entrata(const std::string &data, const std::string &descrizione, int importo): Movimento(data,descrizione,importo) {
+Entrata::Entrata(const std::string &data, const std::string &descrizione, double importo): Movimento(data,descrizione,importo) {
 
     this->record = data + ";" + descrizione + ";" + "Entrata" + ";" + std::to_string(importo);
 }
